requirements.txt
toml
httpx
bleak
uvicorn
fastapi
psycopg2-binary
pytz
pydantic
ollama
openai-whisper
python-multipart